Fix training/validation dataset split
Hi Andrea,

I was looking at your code and it seems to me that the training and validation datasets are assigned incorrectly. For example, using the mESC dataset and default parameters, `n_cells_to_validation = 958` and hence the code assigns 958 batches to the `eval_e_dataset` and only 170 batches to `train_e_dataset`, whereas using the original fraction (17 %) it should be 192 batches for validation and 936 for training accordingly? I added a minimal change to resolve this. Maybe I misunderstood something though, let me know if that's the case!